move and rename ruletype enum
